+++ microROS_test/lib/SPARK_MAX/SPARK_MAX.h
@@ -100,7 +100,7 @@ public:
 
     uint8_t set_control_frame(const control_mode mode, const float setpoint); // Command SPARK MAX output
 
-    void update_status_0(int applied_output); // Updates the applied ouput
+    void update_status_0(float applied_output); // Updates the applied ouput
     void update_status_1(float velocity, float temperature, float voltage, float current); // Updates the Velocity, Temperature, Voltage, and Current
     void update_status_2(float position); // Updates the Position
